+++ example/lib/fullscreen_gallery.dart
@@ -26,6 +26,7 @@ class _FullscreenGalleryState extends State<FullscreenGallery> {
     super.initState();
     controller = GalleryController(
       gallerySetting: GallerySetting(
+          enableCropper: true,
           maximum: 1,
           requestType: RequestType.common,
           onItemClick: (item, list) {},